requirements.txt
# Core dependencies with specific versions
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0

# Optional dependencies
flask==2.3.3
pydantic>=2.0.0
langchain>=0.0.312
langchain_community
tiktoken>=0.4.0
requests>=2.28.0
backoff>=2.0.0
anthropic==0.22.0
httpx>=0.24.1,<1.0.0
python-dotenv>=0.21.0